Replaced AsyncStorage
Installed @react-native-community/async-storage
Replaced imports
Replaced mocks
