+++ app/__e2e__/src/native/async-storage.js
@@ -1,0 +1,15 @@
+let storage = {}
+
+const AsyncStorage = {
+  getItem: async (key) => {
+    return storage[key]
+  },
+  setItem: async (key, value) => {
+    storage[key] = value
+  },
+  clear: async () => {
+    storage = {}
+  },
+}
+
+export default AsyncStorage
