+++ app/__e2e__/src/native/react-native.js
@@ -1,17 +1,3 @@
-let storage = {}
-
-export const AsyncStorage = {
-  getItem: async (key) => {
-    return storage[key]
-  },
-  setItem: async (key, value) => {
-    storage[key] = value
-  },
-  clear: async () => {
-    storage = {}
-  },
-}
-
 export const Linking = {
   addEventListener: () => { },
   removeEventListener: () => { },
