+++ app/android/app/build.gradle
@@ -151,6 +151,7 @@ android {
 }
 
 dependencies {
+    implementation project(':@react-native-community_async-storage')
     compile project(':react-native-svg')
     compile project(':react-native-device-info')
     compile project(':react-native-gesture-handler')
