+++ app/jest.setup.js
@@ -3,7 +3,7 @@ import Adapter from 'enzyme-adapter-react-16'
 
 Enzyme.configure({ adapter: new Adapter() })
 
-jest.mock('AsyncStorage', () => ({
+jest.mock('@react-native-community/async-storage', () => ({
   setItem: jest.fn(),
   getItem: jest.fn(),
   removeItem: jest.fn(),
