+++ app/lib/services/storage.js
@@ -1,4 +1,4 @@
-import { AsyncStorage } from 'react-native'
+import AsyncStorage from '@react-native-community/async-storage'
 
 export async function getAccount () {
   try {
