+++ app/package.json
@@ -22,6 +22,7 @@
   ],
   "license": "Apache-2.0",
   "dependencies": {
+    "@react-native-community/async-storage": "^1.2.4",
     "@types/js-base64": "^2.3.1",
     "axios": "^0.18.0",
     "buffer": "^5.2.1",
